backend/requirements.txt
flask>=3.0
flask-cors>=4.0
psycopg2-binary>=2.9
requests>=2.31
openpyxl>=3.1
python-dotenv>=1.0
google-generativeai>=0.8
